Bug 2051136 - Extend the Lockstore unlock/cache timeout range to 64-bit milliseconds r=simonf
The unlock-KEK timeout and the createKek cache timeout crossed the
IDL / C++ / FFI boundary as uint32_t milliseconds, capping the maximum
window at U32_MAX ms (~49 days). Widen timeoutMs / cacheTimeoutMs to
uint64_t across nsILockstore, LockstoreService and the keystore FFI; the
Rust core already used std::time::Duration, so no core type changed.

Add an unlock_deadline() helper that clamps the requested window to a
~10-year maximum and computes the deadline with Instant::checked_add, so
a large u64 timeout cannot overflow the monotonic clock and panic. It is
used at all three KEK-cache insertion sites.

xpcshell tests cover a timeout of 2**32 ms (which truncates to 0 under
uint32_t, leaving the KEK locked) for both unlockKek and createKek, plus
a Number.MAX_SAFE_INTEGER timeout that would overflow Instant + Duration
without the clamp. A Rust unit test exercises unlock_deadline directly.

Differential Revision: https://phabricator.services.mozilla.com/D309256

+++ security/lockstore/LockstoreService.cpp
@@ -277,7 +277,7 @@ LockstoreService::IsKekUnlocked(const nsACString& aKekRef, bool* aOut) {
 
 nsresult LockstoreService::DoUnlockKek(const nsACString& aKekRef,
                                        const nsACString& aSecret,
-                                       uint32_t aTimeoutMs) {
+                                       uint64_t aTimeoutMs) {
   LOCKSTORE_SYNC_PREAMBLE;
   return keystore_unlock_kek(mKeystore, &aKekRef, &aSecret, aTimeoutMs);
 }
@@ -388,7 +388,7 @@ Result<nsTArray<uint8_t>, nsresult> LockstoreService::DoGetDek(
 
 Result<nsCString, nsresult> LockstoreService::DoCreateKek(
     const nsACString& aKekType, const nsACString& aIdentifier,
-    const nsACString& aSecret, uint32_t aCacheTimeoutMs) {
+    const nsACString& aSecret, uint64_t aCacheTimeoutMs) {
   LOCKSTORE_SYNC_PREAMBLE;
   nsCString out;
   MOZ_TRY(keystore_create_kek(mKeystore, &aKekType, &aIdentifier, &aSecret,
@@ -410,7 +410,7 @@ nsresult LockstoreService::DoDeleteKek(const nsACString& aKekRef) {
 
 NS_IMETHODIMP
 LockstoreService::UnlockKek(const nsACString& aKekRef,
-                            const nsACString& aSecret, uint32_t aTimeoutMs,
+                            const nsACString& aSecret, uint64_t aTimeoutMs,
                             JSContext* aCx, Promise** aPromise) {
   return ImplXpcomMethod(this, aCx, aPromise, &LockstoreService::DoUnlockKek,
                          nsCString{aKekRef}, nsCString{aSecret}, aTimeoutMs);
@@ -535,7 +535,7 @@ LockstoreService::GetDek(const nsACString& aCollection,
 NS_IMETHODIMP
 LockstoreService::CreateKek(const nsACString& aKekType,
                             const nsACString& aIdentifier,
-                            const nsACString& aSecret, uint32_t aCacheTimeoutMs,
+                            const nsACString& aSecret, uint64_t aCacheTimeoutMs,
                             JSContext* aCx, Promise** aPromise) {
   return ImplXpcomMethod(this, aCx, aPromise, &LockstoreService::DoCreateKek,
                          nsCString{aKekType}, nsCString{aIdentifier},
